+++ NEWS.md
@@ -1,8 +1,15 @@
 # EMC2 3.1.1
 
+## New features
+
+* added thin to fit/run_emc which can either be set to TRUE to automatically thin based on ESS, or on a numeric to only keep 1/x samples
+
+
 ## Bug fixes
 
 * Small bug fixes in plot_data to make it more flexible
+
+* cleared up argumentation of run_emc/fit
 
 # EMC2 3.1.0